Recode Maintenance Command - Fix state not loading on server start - Add localised messages - Fix placeholder not getting replaced in kick message
