+++ src/main/kotlin/yv/tils/smp/mods/server/maintenance/MaintenanceCMD.kt
@@ -15,11 +15,7 @@ import yv.tils.smp.utils.internalAPI.Placeholder
 import yv.tils.smp.utils.internalAPI.Vars
 
 class MaintenanceCMD {
-    companion object {
-        var maintenance: Boolean = false
-    }
-
-    private var oldState: Boolean = false
+    val maintenanceHandler = MaintenanceHandler()
 
     val command = commandTree("maintenance") {
         withPermission("yvtils.smp.command.maintenance")
@@ -28,127 +24,12 @@ class MaintenanceCMD {
         stringArgument("state", true) {
             replaceSuggestions(ArgumentSuggestions.strings("true", "false", "toggle"))
             anyExecutor { sender, args ->
-                maintenance(sender, args)
+                maintenanceHandler.maintenance(sender, args)
             }
         }
 
         anyExecutor { sender, _ ->
-            maintenance(sender)
+            maintenanceHandler.maintenance(sender)
         }
-    }
-
-    private fun maintenance(sender: CommandSender, args: CommandArguments? = null) {
-        val state = if (args?.get(0) == null) {
-            "toggle"
-        } else {
-            args[0].toString()
-        }
-
-        when (state) {
-            "toggle" -> {
-                oldState = maintenance
-                maintenance = !maintenance
-            }
-
-            "true" -> {
-                oldState = maintenance
-                maintenance = true
-            }
-
-            "false" -> {
-                oldState = maintenance
-                maintenance = false
-            }
-        }
-
-        globalAnnouncement()
-        senderAnnouncement(sender, state)
-        saveState()
-    }
-
-    private fun globalAnnouncement() {
-        if (oldState == maintenance) {
-            return
-        }
-
-        for (player in Bukkit.getOnlinePlayers()) {
-            if (maintenance) {
-                if (!player.hasPermission("yvtils.smp.bypass.maintenance")) {
-                    player.kick(
-                        Language().getMessage(LangStrings.MAINTENANCE_PLAYER_NOT_ALLOWED_TO_JOIN_KICK_MESSAGE),
-                        PlayerKickEvent.Cause.PLUGIN
-                    )
-                }
-            }
-        }
-    }
-
-    private fun senderAnnouncement(sender: CommandSender, event: String) {
-        if (oldState == maintenance) {
-            sender.sendMessage(
-                Placeholder().replacer(
-                    Language().getMessage(
-                        LangStrings.MAINTENANCE_ALREADY_STATE
-                    ),
-                    listOf("prefix"),
-                    listOf(Vars().prefix)
-                )
-            )
-            return
-        }
-
-        when (event) {
-            "true" -> {
-                sender.sendMessage(
-                    Placeholder().replacer(
-                        Language().getMessage(
-                            LangStrings.MAINTENANCE_COMMAND_ACTIVATE
-                        ),
-                        listOf("prefix"),
-                        listOf(Vars().prefix)
-                    )
-                )
-            }
-
-            "false" -> {
-                sender.sendMessage(
-                    Placeholder().replacer(
-                        Language().getMessage(
-                            LangStrings.MAINTENANCE_COMMAND_DEACTIVATE
-                        ),
-                        listOf("prefix"),
-                        listOf(Vars().prefix)
-                    )
-                )
-            }
-
-            "toggle" -> {
-                if (maintenance) {
-                    sender.sendMessage(
-                        Placeholder().replacer(
-                            Language().getMessage(
-                                LangStrings.MAINTENANCE_COMMAND_ACTIVATE
-                            ),
-                            listOf("prefix"),
-                            listOf(Vars().prefix)
-                        )
-                    )
-                } else {
-                    sender.sendMessage(
-                        Placeholder().replacer(
-                            Language().getMessage(
-                                LangStrings.MAINTENANCE_COMMAND_DEACTIVATE
-                            ),
-                            listOf("prefix"),
-                            listOf(Vars().prefix)
-                        )
-                    )
-                }
-            }
-        }
-    }
-
-    private fun saveState() {
-        Config().changeValue("maintenance", maintenance)
     }
 }
