+++ src/main/kotlin/yv/tils/smp/mods/server/maintenance/PlayerLogin.kt
@@ -1,10 +1,13 @@
 package yv.tils.smp.mods.server.maintenance
 
 import org.bukkit.entity.Player
+import org.bukkit.event.player.PlayerKickEvent
 import org.bukkit.event.player.PlayerLoginEvent
 import yv.tils.smp.utils.configs.global.Config
 import yv.tils.smp.utils.configs.language.LangStrings
 import yv.tils.smp.utils.configs.language.Language
+import yv.tils.smp.utils.internalAPI.Placeholder
+import yv.tils.smp.utils.internalAPI.Vars
 
 class PlayerLogin {
 
@@ -23,9 +26,19 @@ class PlayerLogin {
     }
 
     private fun checkPerm(e: PlayerLoginEvent, player: Player) {
-        if (MaintenanceCMD.maintenance && !player.hasPermission("yvtils.smp.bypass.maintenance")) {
+        if (MaintenanceHandler.maintenance && !player.hasPermission("yvtils.smp.bypass.maintenance")) {
             e.result = PlayerLoginEvent.Result.KICK_FULL
-            e.kickMessage(Language().getMessage(LangStrings.MAINTENANCE_PLAYER_NOT_ALLOWED_TO_JOIN_KICK_MESSAGE))
+            player.kick(
+                Placeholder().replacer(
+                    Language().getMessage(
+                        player.uniqueId,
+                        LangStrings.MAINTENANCE_PLAYER_NOT_ALLOWED_TO_JOIN_KICK_MESSAGE
+                    ),
+                    listOf("prefix"),
+                    listOf(Vars().prefix)
+                ),
+                PlayerKickEvent.Cause.PLUGIN
+            )
         } else {
             funcStarter(state++, e)
         }
