+++ src/main/kotlin/yv/tils/smp/mods/server/motd/GenerateMOTD.kt
@@ -7,6 +7,7 @@ import org.bukkit.Bukkit
 import yv.tils.smp.YVtils
 import yv.tils.smp.mods.admin.vanish.Vanish
 import yv.tils.smp.mods.server.maintenance.MaintenanceCMD
+import yv.tils.smp.mods.server.maintenance.MaintenanceHandler
 import yv.tils.smp.utils.configs.global.Config
 import yv.tils.smp.utils.configs.server.ServerConfig
 import yv.tils.smp.utils.internalAPI.Placeholder
@@ -16,7 +17,7 @@ import java.util.*
 class GenerateMOTD {
     fun onServerPing(e: PaperServerListPingEvent) {
         if (Config.config["modules.server"] as Boolean) {
-            if (MaintenanceCMD.maintenance) {
+            if (MaintenanceHandler.maintenance) {
                 e.version = "Server under maintenance"
                 e.protocolVersion = 0
                 e.maxPlayers = 0
